ontology_toolkit/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ontology-toolkit"
version = "0.3.0"
description = "Инструмент для управления онтологией проекта с опорой на FPF"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "System Career Team"}
]
keywords = ["ontology", "fpf", "knowledge-management", "concepts", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0.1",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "anthropic>=0.34.0",
    "pydantic>=2.8.0",
    "networkx>=3.3",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]
ai-openai = [
    "openai>=1.0.0",
]
ai-gemini = [
    "google-generativeai>=0.3.0",
]
ai-grok = [
    "openai>=1.0.0",
]
ai-all = [
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
]

[project.scripts]
ontology = "ontology_toolkit.cli.main:app"

[tool.setuptools]
packages = ["ontology_toolkit", "ontology_toolkit.core", "ontology_toolkit.cli", "ontology_toolkit.io", "ontology_toolkit.ai", "ontology_toolkit.mcp"]
package-dir = {"ontology_toolkit" = "."}

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
